1. fix a bug in fetchmail.pl readdata() which always returns true    after remote pop3 socket is closed unespectedly, this caused a fever    loop in fetchmail(), thus introduced huge process...    (reported by Mike Fahey, mfahey.AT.enter.net) 2. Romanian (ro & ro.utf8) lang and template files translation update    (thanks to Gabriel Hojda, gabriel.hojda.AT.gmail.com)
git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@179 9f51437f-89f6-0310-a5ff-ff9a3806d2ee

+++ src/cgi-bin/openwebmail/shares/fetchmail.pl
@@ -261,25 +261,27 @@ sub sendcmd {
       alarm 0;
    };
    return 0 if ($@); 		# timeout
+   return 0 if ($ret eq '');	# socket not available?
 
    @{$r_result}=split(/\s+/, $ret);
    shift @{$r_result} if (${$r_result}[0]=~/^[\+\-]/); # rm str +OK or -ERR from @result
    return 1 if ($ret!~/^\-/);
-
    return 0;
 }
 
 sub readdata {
    my ($socket, $r_line, $timeout)=@_;
-
    $timeout=60 if ($timeout<=0);
+
+   ${$r_line}='';	# empty line buff
    eval {
       alarm $timeout; local $SIG{ALRM}= sub {die "alarm\n"};
       ${$r_line}=<$socket>;
       alarm 0;
    };
+   return 0 if ($@); 			# timeout
+   return 0 if (${$r_line} eq '');	# socket not available?
 
-   return 0 if ($@); 		# timeout
    return 1;
 }
